changed a couple API routes (see readme.md for changes) and updated frontend. all protected routes must now have an Authorization header.
